requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==24.2.0
click==8.1.7
Django==4.2.5
django-debug-toolbar==4.3.0
django-filter==23.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.2
pathspec==0.12.1
Pillow==10.0.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
sqlparse==0.4.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
typing_extensions==4.10.0
uritemplate==4.1.1
